Wonf merge

+++ core/include/core/mem.h
@@ -24,5 +24,5 @@ namespace core
 		return sz;
 	}
 
-	void Wide2Char(const WCHAR* data, char* out, size_t len);
+	void Wide2Char(const WCHAR* data, char* out, UINT len);
 }
