+++ core/source/mem.cpp
@@ -49,7 +49,7 @@ namespace core {
 		return API(KERNEL32, HeapFree)(proc_heap, NULL, heap);
 	}
 
-	void Wide2Char(const WCHAR* data, char* out, size_t len)
+	void Wide2Char(const WCHAR* data, char* out, UINT len)
 	{
 		for (int i = 0; i < len; i++) {
 			if (!data[i]) {
